refactor(model): reorganize folders
